Fix follow functionality
