+++ frontend/src/actions/playlist_actions.js
@@ -2,8 +2,7 @@
 import { getPlaylist, 
     getPlaylists, 
     getFollowedPlaylists, 
-    postReaction,
-    postFollow 
+    postReaction
 } from '../util/playlist_api_util';
 import { RECEIVE_REACTIONS } from './reaction_actions';
 
@@ -12,7 +11,6 @@ export const RECEIVE_PLAYLISTS = 'RECEIVE_PLAYLISTS';
 export const RECEIVE_PLAYLIST = 'RECEIVE_PLAYLIST';
 // export const RECEIVE_FOLLOWED_PLAYLISTS = 'RECEIEVE_FOLLOWED_PLAYLISTS';
 export const RECEIVE_REACTION = 'RECEIVE_REACTION';
-export const RECEIVE_FOLLOW = 'RECEIVE_FOLLOW';
 
 export const receivePlaylists = playlists => ({
     type: RECEIVE_PLAYLISTS,
@@ -29,10 +27,6 @@ export const receiveReaction = reaction => ({
     reaction
 });
 
-export const receiveFollow = follow => ({
-    type: RECEIVE_FOLLOW,
-    follow
-});
 // export const receiveFollowedPlaylists = playlists => ({
 //     type: RECEIVE_FOLLOWED_PLAYLISTS,
 //     playlists
@@ -55,12 +49,6 @@ export const sendReaction = (reactionData) => dispatch => (
         .then(res => dispatch(receiveReaction(res.data)))
         .catch(err => console.log(err))
 );
-
-export const sendFollow = followData => dispatch => (
-    postFollow(followData)
-        .then(res => dispatch(receiveFollow(res.data)))
-        .catch(err => console.log(err))
-)
 
 // export const fetchFollowedPlaylists = () => dispatch =>
 //     getFollowedPlaylists(id)
