+++ frontend/src/components/follows/follows_container.js
@@ -1,5 +1,8 @@
 import { connect } from 'react-redux';
-import { fetchFollows, fetchUserFollows } from '../../actions/follow_actions';
+import { fetchUserFollows, 
+  sendFollow, 
+  deleteFollow 
+} from '../../actions/follow_actions';
 
 import Follow from './follows';
 
@@ -11,8 +14,9 @@ const mapStateToProps = state => {
 
 const mapDispatchToProps = dispatch => {
   return {
-    fetchFollows: () => dispatch(fetchFollows()),
-    fetchUserFollows: (id) => dispatch(fetchUserFollows(id))
+    fetchUserFollows: (id) => dispatch(fetchUserFollows(id)),
+    sendFollow: (followData) => dispatch(sendFollow(followData)),
+    deleteFollow: (followId) => dispatch(deleteFollow(followId))
   };
 };
