+++ frontend/src/components/profile/profile_container.js
@@ -2,20 +2,16 @@ import { connect } from "react-redux";
 import { withRouter } from 'react-router-dom';
 
 import Profile from './profile';
-import * as FollowActions from '../../actions/follow_actions';
 import * as SessionActions from '../../actions/session_actions';
 
 const mapStateToProps = (state, ownProps) => {
     return {
-        follows: state.entities.follows,
         user: state.session.user
     };
 };
 
 const mapDispatchToProps = dispatch => {
     return {
-        fetchFollows: () => dispatch(FollowActions.fetchFollows()),
-        deleteFollow: followId => dispatch(FollowActions.deleteFollow(followId)),
         fetchUser: userId => dispatch(SessionActions.fetchUser(userId))
     };
 };
