+++ frontend/src/reducers/follows_reducer.js
@@ -1,6 +1,6 @@
 import {
-    RECEIVE_FOLLOWS,
     RECEIVE_USER_FOLLOWS,
+    RECEIVE_FOLLOW,
     REMOVE_FOLLOW
 } from '../actions/follow_actions';
 
@@ -9,16 +9,16 @@ const FollowsReducer = (state = { all: {}, user: {}}, action) => {
     let newState = Object.assign({}, state);
 
     switch(action.type) {
-        case RECEIVE_FOLLOWS:
-            newState.all = action.follows.data;
-            return newState;
-
         case RECEIVE_USER_FOLLOWS:
             newState.user = action.follows.data;
             return newState;
 
         case REMOVE_FOLLOW:
             delete newState[action.followId];
+            return newState;
+
+        case RECEIVE_FOLLOW:
+            newState.follow = action.follow;
             return newState;
             
         default:
