+++ frontend/src/reducers/playlists_reducer.js
@@ -2,8 +2,7 @@
 import { RECEIVE_PLAYLISTS, 
     RECEIVE_PLAYLIST, 
     RECEIVE_FOLLOWED_PLAYLISTS, 
-    RECEIVE_REACTION,
-    RECEIVE_FOLLOW
+    RECEIVE_REACTION
 } from '../actions/playlist_actions';
 
 
@@ -22,9 +21,6 @@ const PlaylistsReducer = (state = { all: [], currentPlaylist: {}, reactions: [] 
         //     return newState;
         case RECEIVE_REACTION:
             newState.reaction = action.reaction;
-            return newState;
-        case RECEIVE_FOLLOW:
-            newState.follow = action.follow;
             return newState;
         default:
             return state;
